prototypes/requirements.txt
sentence-transformers>=2.2.2
scikit-learn>=1.0.2
pandas>=1.3.5
matplotlib>=3.5.1
plotly>=5.6.0
networkx>=2.6.3
nltk>=3.6.7
jupyter>=1.0.0
umap-learn>=0.5.3
